+++ python/eyes_selenium/CHANGELOG.md
@@ -1,8 +1,12 @@
 ## [vNext]
+### Added
+- Region Id support for coded regions [Trello 47](https://trello.com/c/pv51sCYA)
 ### Fixed
 - Log Server.log notification with requested level [Trello 2953](https://trello.com/c/IFCsr37o)
 - Cleanup obsolete selector type conversion, fix repr for MobileBy TargetPaths [Trello 2955](https://trello.com/c/YDPO9isU)
 - Missing license file in distributed packages [Trello 2962](https://trello.com/c/0HdBnfAK)
+- Layout padding not appearing [Trello 2909](https://trello.com/c/ZzUbo9ir)
+- UFG Eyes raise exception when render browser is not specified
 
 ## [5.7.0] - 2022-06-15
 ### Added